refactor(core): keep fastapi out of worker/scheduler import closures (B1-B3)
B1: DbSession/OptionalDbSession move from src.core.database to the new
API-only src.core.db_deps (fastapi Depends lives there); database.py is
now fastapi-free. All import sites updated mechanically.

B2: UserPrincipal moves from src.core.auth to the new fastapi-free
src.core.principal; auth.py imports it from there. Worker/scheduler
closure modules (repositories.executions, org_filter, agent_executor,
mcp_server.tools.execution) plus routers/tests now import from principal.

B3: pubsub.py gains `from __future__ import annotations` and moves the
fastapi WebSocket import under TYPE_CHECKING — all uses are annotations.

B4 (solutions guard) is NOT APPLICABLE on this branch: the plan was
verified against the unmerged solutions-success-criteria worktree;
api/src/services/solutions/guard.py does not exist on main and
get_session_factory() imports no guard here.

Result: fastapi/starlette gone from both src.worker.app and
src.scheduler.main closures (guardrail output).
pyright full-api: 42 errors, all pre-existing host reportMissingImports
(0 non-missing-import errors). ruff clean. Unit suite: 4068 passed,
2 expected-red guardrail failures (pending Tasks 4-6 heavies).
