+++ api/src/core/auth.py
@@ -6,14 +6,15 @@ Supports JWT bearer token authentication with user context injection.
 """
 
 import logging
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import TYPE_CHECKING, Annotated
 from uuid import UUID
 
 from fastapi import Depends, HTTPException, Request, status
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
 
-from src.core.database import DbSession
+from src.core.db_deps import DbSession
+from src.core.principal import UserPrincipal
 from src.core.security import decode_token
 from shared.role_cache import get_user_roles
 
@@ -24,58 +25,6 @@ logger = logging.getLogger(__name__)
 
 # HTTP Bearer token scheme
 bearer_scheme = HTTPBearer(auto_error=False)
-
-
-@dataclass
-class UserPrincipal:
-    """
-    Authenticated user principal.
-
-    Represents an authenticated user with their identity and permissions.
-    All user info is extracted from JWT claims (no database lookup required).
-
-    Auth model:
-    - is_superuser=true, org_id=UUID: Platform admin in an org
-    - is_superuser=false, org_id=UUID: Regular org user
-    - is_superuser=true, org_id=None: System account (global scope)
-    - is_superuser=false, org_id=None: INVALID (rejected at token parsing)
-    """
-    user_id: UUID
-    email: str
-    organization_id: UUID | None  # User's org (None for system accounts)
-    name: str = ""
-    is_active: bool = True
-    is_superuser: bool = False
-    is_verified: bool = False
-    roles: list[str] = field(default_factory=list)
-    # Role identity used by table-policy `has_role` evaluator. Populated by
-    # `get_execution_context` from the `user_roles` table; empty for token-only
-    # principals (e.g. system accounts) and embed sessions.
-    role_ids: list[UUID] = field(default_factory=list)
-    role_names: list[str] = field(default_factory=list)
-    embed: bool = False  # True for embed session tokens (scoped to app_id)
-    jti: str | None = None  # JWT ID for embed tokens (used for execution scoping)
-    app_id: str | None = None  # App ID for embed tokens
-    form_id: str | None = None  # Form ID for form embed tokens
-    verified_params: dict[str, str] | None = None  # HMAC-verified query params
-
-    @property
-    def is_platform_admin(self) -> bool:
-        """Check if user is a platform admin (superuser)."""
-        return self.is_superuser
-
-    @property
-    def is_system_account(self) -> bool:
-        """Check if this is a system account (superuser with no org)."""
-        return self.is_superuser and self.organization_id is None
-
-    def has_role(self, role: str) -> bool:
-        """Check if user has a specific role."""
-        return role in self.roles
-
-    def has_any_role(self, *roles: str) -> bool:
-        """Check if user has any of the specified roles."""
-        return any(role in self.roles for role in roles)
 
 
 @dataclass
