+++ api/src/routers/ai_pricing.py
@@ -20,7 +20,7 @@ from sqlalchemy import select
 
 from src.core.auth import CurrentActiveUser, RequirePlatformAdmin
 from src.core.cache import get_shared_redis
-from src.core.database import DbSession
+from src.core.db_deps import DbSession
 from src.core.log_safety import log_safe
 from src.models import (
     AIModelPricingCreate,
