+++ api/src/routers/export_import.py
@@ -19,7 +19,7 @@ from sqlalchemy import select
 from sqlalchemy.orm import selectinload
 
 from src.core.auth import CurrentSuperuser
-from src.core.database import DbSession
+from src.core.db_deps import DbSession
 from shared.policies.probe import make_seed_admin_bypass
 from src.core.security import decrypt_with_key, encrypt_secret
 from src.models.enums import ConfigType
