+++ api/src/routers/llm_config.py
@@ -10,7 +10,7 @@ import logging
 from fastapi import APIRouter, HTTPException, status
 
 from src.core.auth import CurrentActiveUser, RequirePlatformAdmin
-from src.core.database import DbSession
+from src.core.db_deps import DbSession
 from src.core.log_safety import log_safe
 from src.models.contracts.llm import (
     EmbeddingConfigRequest,
