+++ api/src/routers/mfa.py
@@ -16,7 +16,7 @@ from pydantic import BaseModel, Field
 from src.core.cache import get_shared_redis
 from src.core.cache.keys import refresh_token_jti_key, TTL_REFRESH_TOKEN
 from src.core.auth import CurrentActiveUser, get_current_user_from_db
-from src.core.database import DbSession
+from src.core.db_deps import DbSession
 from src.core.log_safety import log_safe
 from src.core.security import (
     create_access_token,
