+++ api/src/routers/oauth_sso.py
@@ -25,7 +25,7 @@ from src.core.cache.keys import (
 from src.core.log_safety import log_safe
 from src.config import get_settings
 from src.core.auth import CurrentActiveUser, get_current_user_from_db
-from src.core.database import DbSession
+from src.core.db_deps import DbSession
 from src.core.security import create_access_token, create_refresh_token, generate_csrf_token
 from src.services.oauth_sso import OAuthError, OAuthService
 from src.services.user_provisioning import ensure_user_provisioned, get_user_roles
