+++ api/src/routers/tools.py
@@ -11,7 +11,7 @@ from typing import Literal
 from fastapi import APIRouter, Query
 
 from src.core.auth import CurrentActiveUser
-from src.core.database import DbSession
+from src.core.db_deps import DbSession
 from src.core.org_filter import resolve_org_filter
 from src.models.contracts.agents import ToolInfo, ToolsResponse
 from src.repositories.workflows import WorkflowRepository
