+++ api/src/routers/users.py
@@ -14,7 +14,7 @@ from sqlalchemy import select
 
 from src.config import get_settings
 from src.core.auth import CurrentSuperuser
-from src.core.database import DbSession
+from src.core.db_deps import DbSession
 from src.core.log_safety import log_safe
 from src.core.org_filter import resolve_org_filter, OrgFilterType
 from src.services.audit import emit_audit
